pyproject.toml
[tool.poetry]
name = "sukkolake"
version = "0.1.0"
description = ""
authors = ["a.cheremushkin <a.ch3re359@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
python-dotenv = "^0.21.0"
django-bootstrap4 = "^22.3"
pillow = "^9.3.0"
gunicorn = "^20.1.0"
dj-database-url = "^1.2.0"
psycopg2 = "^2.9.5"
whitenoise = "^6.2.0"


[tool.poetry.group.dev.dependencies]
loguru = "^0.6.0"
flake8 = "^6.0.0"
isort = "^5.11.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
